Moved ArmyRepository to subpackage
 Changes to be committed:
	renamed:    src/main/java/com/ardaslegends/repository/ArmyRepository.java -> src/main/java/com/ardaslegends/repository/war/army/ArmyRepository.java
	modified:   src/main/java/com/ardaslegends/service/ArmyService.java
	modified:   src/main/java/com/ardaslegends/service/MovementService.java
	modified:   src/main/java/com/ardaslegends/service/time/ScheduleService.java
	modified:   src/test/java/com/ardaslegends/repository/ArmyRepositoryTest.java
	modified:   src/test/java/com/ardaslegends/repository/MovementRepositoryTest.java
	modified:   src/test/java/com/ardaslegends/service/ArmyServiceTest.java
	modified:   src/test/java/com/ardaslegends/service/MovementServiceTest.java
	modified:   src/test/java/com/ardaslegends/service/ScheduleServiceTest.java
