+++ src/test/java/com/ardaslegends/repository/MovementRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.ardaslegends.repository;
 
 import com.ardaslegends.domain.*;
+import com.ardaslegends.repository.war.army.ArmyRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
